Alpha 1.0  Changes to be committed: 	modified:   README.md 	modified:   build.sh 	modified:   resources/fdisk-command 	deleted:    resources/init_resize 	modified:   scripts/boot.sh 	modified:   scripts/environment.sh 	modified:   scripts/final.sh 	modified:   scripts/image.sh 	modified:   scripts/kernel.sh 	modified:   scripts/readme.md 	modified:   scripts/rootfs.sh

+++ scripts/image.sh
@@ -1,24 +1,41 @@
+if [ `id -u` -ne 0 ];then
+    echo "Please run this script with 'sudo' or by root user."
+    exit 1
+fi
 if test -z "$whereisme"
 then
-  echo "Found null path value. please fix the path value before running this script. See ./README.md and ./scripts/readme.md for more infomation."
+  echo "Found null path value. please fix the path value before running this script. Check ./README.md and ./scripts/readme.md to get more infomation."
   exit 1
 fi
 echo "Parpareing the empty system image(3000M)..."
-dd if=/dev/zero of=$whereisme/image bs=1M count=3000
-echo "Setup the empty image..."
-fdisk image < $whereisme/resources/fdisk-command
-sudo -av kpartx $whereisme/image
-echo "Making vfat file system for boot partition..."
-mkfs.vfat /dev/mapper/$(ls /dev/mapper | grep p1)
-echo "Making ext4 file system for rootfs partition..."
-mkfs.ext4 /dev/mapper/$(ls /dev/mapper | grep p2)
-echo "Mounting the image..."
-rm -rf $whereisme/rootfs
-rm -rf $whereisme/boot
-mkdir $whereisme/rootfs
-mkdir $whereisme/boot
-mount /dev/mapper/$(ls /dev/mapper | grep p2) $whereisme/rootfs
-mount /dev/mapper/$(ls /dev/mapper | grep p1) $whereisme/boot
+if [ -f "$whereisme/image" ]; then
+  rm $whereisme/image
+fi
 
+dd if=/dev/zero of=$whereisme/image bs=1M count=3000
+echo "Setup the image file..."
+fdisk $whereisme/image < $whereisme/resources/fdisk-command
+losetup --offset 1048576 --sizelimit 103809536 /dev/loop6 $whereisme/image
+if [ $? -ne 0 ]; then
+    echo "Device path /dev/loop6 is unavailable. Check the paragraph about image.sh in ./scripts/readme.md to get more imfomation."
+    exit 1
+fi
+
+losetup --offset 105906176 /dev/loop7 $whereisme/image
+if [ $? -ne 0 ]; then
+    echo "Device path /dev/loop7 is unavailable. Check the paragraph about image.sh in ./scripts/readme.md to get more imfomation."
+    exit 1
+fi
+echo "Making vfat file system for boot partition..."
+mkfs.vfat /dev/loop6
+echo "Making ext4 file system for rootfs partition..."
+mkfs.ext4 /dev/loop7
+echo "Mounting the image..."
+rm -rf $whereisme/boot
+rm -rf $whereisme/rootfs
+mkdir $whereisme/boot
+mkdir $whereisme/rootfs
+mount /dev/loop6 $whereisme/boot
+mount /dev/loop7 $whereisme/rootfs
